Move `TestApp` into a submodule
